Backend_Python/requirements.txt
flask
pdfplumber
requests
pymongo
python-dotenv